cached regions and zones. interfaces, structs extracted into a separate file
